Move efi_main into efi/loader
Move the efi_main routine out of libefi into sys/boot/efi/loader.
Since boot1 has its own efi_main routine, this effectively prevents
boot1 from linking with libefi. By moving it out, we can share code
better (though though some refactoring with boot1's efi_main and
loader.efi's efi_main is definitely in order).

Sponsored by: Netflix


Former-commit-id: fd5917af4e546dd2044e155b49d44d94d6099504

+++ sys/boot/efi/include/efidevp.h
@@ -141,7 +141,7 @@ name|DevicePathNodeLength
 parameter_list|(
 name|a
 parameter_list|)
-value|( ((a)->Length[0]) | ((a)->Length[1]<< 8) )
+value|((size_t)(((a)->Length[0]) | ((a)->Length[1]<< 8)))
 end_define
 
 begin_define
